refactor: move enum-related entities to a dedicated `enums` module
- Relocated `content_source_status_enum`, `ingestion_job_status_enum`, and `source_type_enum_entity` to `src/domain/entities/enums`.
- Updated imports across the project to reflect the new module structure.
