+++ src/domain/entities/chunk_entity.py
@@ -4,7 +4,7 @@ from uuid import UUID, uuid4
 
 from pydantic import BaseModel, Field
 
-from src.domain.entities.source_type_enum_entity import SourceType
+from src.domain.entities.enums.source_type_enum_entity import SourceType
 
 
 class ChunkEntity(BaseModel):
