+++ src/domain/mappers/chunk_mapper.py
@@ -9,7 +9,7 @@ from uuid import uuid4
 
 from langchain_core.documents import Document
 from src.domain.entities.chunk_entity import ChunkEntity
-from src.domain.entities.source_type_enum_entity import SourceType
+from src.domain.entities.enums.source_type_enum_entity import SourceType
 from src.infrastructure.repositories.vector.models.chunk_model import ChunkModel
