+++ tests/vector/test_youtube_weaviate_service.py
@@ -2,7 +2,7 @@ from uuid import uuid4
 
 import pytest
 from src.domain.entities.chunk_entity import ChunkEntity
-from src.domain.entities.source_type_enum_entity import SourceType
+from src.domain.entities.enums.source_type_enum_entity import SourceType
 from src.infrastructure.repositories.vector.models.chunk_model import ChunkModel
 from src.infrastructure.services.youtube_vector_service import YouTubeVectorService
